server/requirements.txt
aiopg==1.4.0
asgiref==3.7.2
async-timeout==4.0.3
asyncio==3.4.3
asyncssh==2.14.2
attrs==23.2.0
autobahn==23.6.2
Automat==22.10.0
bidict==0.23.1
certifi==2024.2.2
cffi==1.16.0
channels==4.0.0
channels_postgres==1.0.4
charset-normalizer==3.3.2
chess==1.10.0
constantly==23.10.4
cryptography==42.0.5
daphne==4.1.2
Django==5.0.4
django-cors-headers==4.3.1
django-extensions==3.2.3
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
h11==0.14.0
hyperlink==21.0.0
idna==3.7
incremental==22.10.0
msgpack==1.0.8
oauthlib==3.2.2
psycopg2-binary==2.9.9
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
PyJWT==2.8.0
pyOpenSSL==24.1.0
python-chess==1.999
python-engineio==4.9.0
requests==2.31.0
requests-oauthlib==2.0.0
rest-framework-simplejwt==0.0.2
schedule==1.2.2
service-identity==24.1.0
simple-websocket==1.0.0
six==1.16.0
sqlparse==0.4.4
Twisted==24.3.0
txaio==23.1.1
typing_extensions==4.11.0
urllib3==2.2.1
uWSGI==2.0.24
wsproto==1.2.0
zope.interface==6.3
